Refactored the ros service wrapper
