--- src/services/ros/types.ts
@@ -1,2 +1,0 @@
-export * from './ros-authentication';
-export * from './ros-realms';
