+++ src/ui/server-administration/users/UsersTableContainer.tsx
@@ -2,14 +2,7 @@ import * as electron from 'electron';
 import * as React from 'react';
 import * as Realm from 'realm';
 
-import {
-  createUser,
-  IRealmFile,
-  IUser,
-  IUserMetadataRow,
-  RealmLoadingMode,
-  updateUserPassword,
-} from '../../../services/ros';
+import * as ros from '../../../services/ros';
 import { showError } from '../../reusable/errors';
 import { ILoadingProgress } from '../../reusable/loading-overlay';
 import {
@@ -27,7 +20,7 @@ export interface IUsersTableContainerState extends IRealmLoadingComponentState {
   isChangePasswordOpen: boolean;
   isCreateUserOpen: boolean;
   selectedUserId: string | null;
-  users: Realm.Results<IUser> | null;
+  users: Realm.Results<ros.IUser> | null;
   progress: ILoadingProgress;
 }
 
@@ -80,19 +73,19 @@ export class UsersTableContainer extends RealmLoadingComponent<
     });
   };
 
-  public getUser = (index: number): IUser | null => {
+  public getUser = (index: number): ros.IUser | null => {
     return this.state.users ? this.state.users[index] : null;
   };
 
-  public getUserFromId = (userId: string): IUser | null => {
-    return this.realm.objectForPrimaryKey<IUser>('User', userId);
+  public getUserFromId = (userId: string): ros.IUser | null => {
+    return this.realm.objectForPrimaryKey<ros.IUser>('User', userId);
   };
 
-  public getUsersRealms = (userId: string): IRealmFile[] => {
-    const user = this.realm.objectForPrimaryKey<IRealmFile>('User', userId);
+  public getUsersRealms = (userId: string): ros.IRealmFile[] => {
+    const user = this.realm.objectForPrimaryKey<ros.IRealmFile>('User', userId);
     if (user) {
       const realms = this.realm
-        .objects<IRealmFile>('RealmFile')
+        .objects<ros.IRealmFile>('RealmFile')
         .filtered('owner = $0', user);
       return realms.slice();
     } else {
@@ -100,7 +93,7 @@ export class UsersTableContainer extends RealmLoadingComponent<
     }
   };
 
-  public getUsersMetadatas = (userId: string): IUserMetadataRow[] => {
+  public getUsersMetadatas = (userId: string): ros.IUserMetadataRow[] => {
     const user = this.getUserFromId(userId);
     return user ? user.metadata.slice() : [];
   };
@@ -125,7 +118,11 @@ export class UsersTableContainer extends RealmLoadingComponent<
   };
 
   public onUserCreated = async (username: string, password: string) => {
-    const userId = await createUser(this.props.user.server, username, password);
+    const userId = await ros.users.create(
+      this.props.user.server,
+      username,
+      password,
+    );
     this.onUserSelected(userId);
   };
 
@@ -147,7 +144,7 @@ export class UsersTableContainer extends RealmLoadingComponent<
     if (user) {
       this.realm.write(() => {
         const metadataRow = this.realm.create<
-          IUserMetadataRow
+          ros.IUserMetadataRow
         >('UserMetadataRow', {
           key: '',
           value: '',
@@ -190,14 +187,14 @@ export class UsersTableContainer extends RealmLoadingComponent<
   };
 
   public onUserPasswordChanged = (userId: string, password: string) => {
-    const success = updateUserPassword(this.props.user, userId, password);
+    const success = ros.users.updatePassword(this.props.user, userId, password);
     if (!success) {
       showError("Couldn't update password");
     }
   };
 
   public onUserRoleChanged = (userId: string, role: UserRole) => {
-    const user = this.realm.objectForPrimaryKey<IUser>('User', userId);
+    const user = this.realm.objectForPrimaryKey<ros.IUser>('User', userId);
     if (user) {
       this.realm.write(() => {
         user.isAdmin = role === UserRole.Administrator;
@@ -210,7 +207,7 @@ export class UsersTableContainer extends RealmLoadingComponent<
   protected gotUser(user: Realm.Sync.User) {
     this.loadRealm({
       authentication: this.props.user,
-      mode: RealmLoadingMode.Synced,
+      mode: ros.realms.RealmLoadingMode.Synced,
       path: '__admin',
       validateCertificates: true,
     });
@@ -223,7 +220,7 @@ export class UsersTableContainer extends RealmLoadingComponent<
   protected onRealmLoaded = () => {
     // Get the users and save them in the state
     this.setState({
-      users: this.realm.objects<IUser>('User').sorted('userId'),
+      users: this.realm.objects<ros.IUser>('User').sorted('userId'),
     });
   };
